% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph.R
\name{add_content_subgraph}
\alias{add_content_subgraph}
\title{Add the content inside the subgraph}
\usage{
add_content_subgraph(text = "", name = "")
}
\arguments{
\item{text}{text}

\item{name}{name}
}
\description{
Add the content inside the subgraph
}
