% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ming_kings.R
\docType{data}
\name{ming_kings}
\alias{ming_kings}
\title{A dataset containing the information about Ming Dynasty Emperors of China}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 16 rows and 2 columns.}
\usage{
ming_kings
}
\description{
A dataset containing the information about Ming Dynasty Emperors of China
}
\keyword{datasets}
